// src/api/model/timeTrackerDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */

export interface TimeTrackerDto {
	id?: string;
	userId?: string;
	workspaceId?: string;
	startedAt?: string;
	/** @nullable */
	endedAt?: string | null;
	/** @pattern ^-?(\d+\.)?\d{2}:\d{2}:\d{2}(\.\d{1,7})?$ */
	trackedDuration?: string;
	isPaused?: boolean;
	/** @nullable */
	lastPausedAt?: string | null;
	subject?: string;
	/** @nullable */
	description?: string | null;
}